Move Doctrine MongoDB ODM Adapter to package
