+++ lib/Adapter/DoctrineMongoDBODM/composer.json
@@ -1,0 +1,23 @@
+{
+    "name": "pagerfanta/doctrine-mongodb-odm-adapter",
+    "type": "library",
+    "description": "Pagerfanta adapter for Doctrine MongoDB ODM",
+    "keywords": ["pagerfanta", "doctrine", "mongodb", "odm"],
+    "license": "MIT",
+    "require": {
+        "php": "^7.2",
+        "doctrine/mongodb-odm": "^2.0",
+        "pagerfanta/core": "^2.4"
+    },
+    "autoload": {
+        "psr-4": {
+            "Pagerfanta\\DoctrineMongoDBODM\\": "./"
+        }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-2.x": "2.x-dev"
+        }
+    },
+    "minimum-stability": "dev"
+}
